refactor: 드롭다운 리팩터링
- 카테고리, 무드를 분리
