+++ client/src/app/components/studio/StudioDropdown.tsx
@@ -134,106 +134,4 @@ const StudioDropdown = Object.assign(DropdownWrapper, {
   Item: DropdownItem,
 });
 
-type CategoryTestProps = {
-  setData: (data: string) => void;
-  componentType: 'category' | 'mood';
-  placeHolder: string;
-  data: string;
-};
-
-const categories = Object.values(CONTENTS_CATEGORY);
-const moods = Object.values(MOODS_CATEGORY);
-
-export const StudioDropdownRenderer = ({ setData, data, componentType, placeHolder }: CategoryTestProps) => {
-  const [selectedKey, setSelectedKey] = useState<string | null>('');
-  const [selectedCode, setSelectedCode] = useState<ContentsCategoryKey>('talk');
-  const [selectedMoodCode, setSelectedMoodCode] = useState<MoodsCategoryKey>('unknown');
-
-  const selectCategory = (key: string) => {
-    setSelectedKey(key);
-    setData(key);
-  };
-
-  useEffect(() => {
-    console.log(data);
-    setSelectedKey(data);
-    if (componentType === 'category') {
-      setSelectedCode(categories.find((c) => c.NAME === data)?.CODE || 'talk');
-    } else {
-      setSelectedMoodCode(moods.find((m) => m.NAME === data)?.CODE || 'unknown');
-    }
-  }, [data]);
-
-  return (
-    <StudioDropdown>
-      {({ inputRef, inputValue, isFocused, handleChangeInput, handleFocusInput, blurDropdown }) => (
-        <>
-          <StudioDropdown.Search
-            ref={inputRef}
-            value={inputValue}
-            onChange={handleChangeInput}
-            onFocus={handleFocusInput}
-            placeholder={placeHolder}
-          />
-          {isFocused && (
-            <StudioDropdown.List>
-              {componentType === 'category' ? (
-                <>
-                  {categories
-                    .filter((c) => c.NAME.includes(inputValue))
-                    .map((c, idx) => (
-                      <StudioDropdown.Item
-                        key={idx}
-                        onClick={() => {
-                          selectCategory(c.NAME);
-                          blurDropdown();
-                          setSelectedCode(c.CODE);
-                        }}
-                      >
-                        <span>{c.NAME}</span>
-                      </StudioDropdown.Item>
-                    ))}
-                </>
-              ) : (
-                <>
-                  {moods
-                    .filter((m) => m.NAME.includes(inputValue))
-                    .map((m, idx) => (
-                      <StudioDropdown.Item
-                        key={idx}
-                        onClick={() => {
-                          selectCategory(m.NAME);
-                          blurDropdown();
-                          setSelectedMoodCode(m.CODE);
-                        }}
-                      >
-                        <span>{m.NAME}</span>
-                      </StudioDropdown.Item>
-                    ))}
-                </>
-              )}
-            </StudioDropdown.List>
-          )}
-          {selectedKey && (
-            <div className="mt-5 flex justify-center">
-              {componentType === 'category' ? (
-                <div className="flex items-center justify-center">
-                  <StudioCategoryCard code={selectedCode} title={selectedKey} />
-                </div>
-              ) : (
-                <div className="relative mt-2 h-10 w-full shadow-md">
-                  {MoodsCategoryPalette({ code: selectedMoodCode })}
-                  <div className="text-content-neutral-inverse absolute left-[45%] top-[20%]">
-                    {moods.find((m) => m.NAME === selectedKey)?.NAME}
-                  </div>
-                </div>
-              )}
-            </div>
-          )}
-        </>
-      )}
-    </StudioDropdown>
-  );
-};
-
 export default StudioDropdown;
